requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyodbc==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0